backend/requirements.txt
# Django and REST Framework
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.0

# PDF Generation
WeasyPrint==60.2

# Task Queue (optional)
celery==5.3.4
redis==5.0.1

# Email
django-anymail==10.2

# Validation
pydantic==2.5.3

# Development and Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
black==23.12.1
isort==5.13.2
flake8==7.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Security
django-filter==23.5 